+++ tiles/src/main/java/com/protomaps/basemap/names/NeNames.java
@@ -2,6 +2,8 @@ package com.protomaps.basemap.names;
 
 import com.onthegomap.planetiler.FeatureCollector;
 import com.onthegomap.planetiler.reader.SourceFeature;
+import com.protomaps.basemap.text.FontRegistry;
+import com.protomaps.basemap.text.TextEngine;
 import java.util.Map;
 
 public class NeNames {
@@ -10,12 +12,38 @@ public class NeNames {
 
   public static FeatureCollector.Feature setNeNames(FeatureCollector.Feature feature, SourceFeature sf,
     int minZoom) {
+    FontRegistry fontRegistry = FontRegistry.getInstance();
+
     for (Map.Entry<String, Object> tag : sf.tags().entrySet()) {
-      var key = tag.getKey();
+      String key = tag.getKey().toString();
+      if (sf.getTag(key) == null) {
+        continue;
+      }
+      String value = sf.getTag(key).toString();
+      var script = Script.getScript(value);
+
+      if (key.startsWith("name_")) {
+        key = key.replace("_", ":");
+      }
+
       if (key.equals("name")) {
-        feature.setAttrWithMinzoom(key, sf.getTag(key), minZoom);
-      } else if (key.startsWith("name_")) {
-        feature.setAttrWithMinzoom(key.replace("_", ":"), sf.getTag(key), minZoom);
+        feature.setAttrWithMinzoom("name", value, minZoom);
+
+        if (!script.equals("Latin") && !script.equals("Generic")) {
+          feature.setAttrWithMinzoom("pmap:script", script, minZoom);
+        }
+
+        String encodedValue = TextEngine.encodeRegisteredScripts(value);
+        feature.setAttrWithMinzoom("pmap:pgf:name", encodedValue, minZoom);
+      }
+
+      if (key.startsWith("name:")) {
+        feature.setAttrWithMinzoom(key, value, minZoom);
+
+        if (fontRegistry.getScripts().contains(script)) {
+          String encodedValue = TextEngine.encodeRegisteredScripts(value);
+          feature.setAttrWithMinzoom("pmap:pgf:" + key, encodedValue, minZoom);
+        }
       }
     }